pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "drako"
version = "2.1.0"
description = "The Python Governance Platform for AI Agents — 60 policy rules, per-tool interception, HITL, FinOps, A2A."
readme = "README.md"
license = "BUSL-1.1"
requires-python = ">=3.10"
authors = [
    { name = "Drako", email = "hello@useagentmesh.com" },
]
keywords = ["ai-governance", "agent-security", "compliance", "eu-ai-act", "langraph", "crewai", "autogen", "dlp", "audit-trail", "mcp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Security",
]
dependencies = [
    "httpx>=0.27.0",
    "click>=8.1.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
crewai = ["crewai>=0.70.0"]
langgraph = ["langgraph>=0.2.0"]
autogen = ["pyautogen>=0.4.0"]
proxy = ["fastapi>=0.110.0", "uvicorn>=0.29.0"]
all = ["crewai>=0.70.0", "langgraph>=0.2.0", "pyautogen>=0.4.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "respx>=0.22.0",
    "coverage>=7.0",
    "hypothesis>=6.100.0",
]

[project.scripts]
drako = "drako.cli.main:cli"

[project.urls]
Homepage = "https://useagentmesh.com"
Documentation = "https://docs.useagentmesh.com"
Repository = "https://github.com/agentmesh/drako-sdk"
Issues = "https://github.com/agentmesh/drako-sdk/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/drako"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py310"
line-length = 100
